+++ lib/dawn/kb/cve_2007_6077.rb
@@ -12,6 +12,8 @@
             :cvss=>"AV:N/AC:M/Au:N/C:P/I:P/A:P",
             :release_date => Date.new(2007, 11, 21),
             :cwe=>"362",
+            :severity=>:high,
+            :priority=>:high,
             :owasp=>"A9", 
             :applies=>["rails"],
             :kind=>Dawn::KnowledgeBase::DEPENDENCY_CHECK,