+++ spec/spec_helper.rb
@@ -1,11 +1,12 @@
 require 'coveralls'
 require 'dawnscanner'
-require 'codesake-commons'
 
 def give_a_number
   require 'securerandom'
   return SecureRandom.random_number(9999)
 end
 
-$logger  = Codesake::Commons::Logging.instance
+require 'logger'
+$logger = Logger.new(STDOUT)
+$logger.datetime_format = '%Y-%m-%d %H:%M:%S'
 Coveralls.wear!